+++ TensorShaderTest/Initializers/ComplexNormalTest.cs
@@ -26,7 +26,7 @@ namespace TensorShaderTest.Initializers {
 
                     Tensor y = Tensor.ComplexDense(x, w);
 
-                    float variance = (float)Tensor.Average(y * y).State;
+                    float variance = Tensor.Average(y * y).State;
 
                     variance_list.Add(variance);
                 }
